Event loop library.
The event loop library serves as an infrastructure for writing
event driven porgrams.
This style of programming is also known as CPS (continuation
passing style), where the flow functions all implement tail position
calls (where the last thing a function does is a call to another function).
When using the event library - instead of calling a function, an event
is added to the event loop with the next flow function to be executed.